refactor(textlnt): add textlint-core class
 - export core to textlint-core as a class
 - textlint is a instance for textlint-core
 - textlint is a singleton object
